requirements.txt
flask>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
scikit-learn>=1.4.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
